fix: vuex-modules 학습
