+++ todo/src/components/TodoList.vue
@@ -18,7 +18,7 @@ export default {
     },
     computed: {
         todos() {
-            return this.$store.state.todos;
+            return this.$store.state.todo.todos;
         }
     },
     methods: {
